requirements.txt
pandas
scikit-learn
ipython
google-genai
